turn subscribe button into component and add to post page
